+++ Bootstrap.php
@@ -27,7 +27,7 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
 
     public function getVersion()
     {
-        return '1.2.3';
+        return '1.2.4';
     }
 
     public function getInfo()
@@ -191,6 +191,8 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
                 );
                 break;
             case '1.2.2':
+                break;
+            case '1.2.3':
                 break;
             default:
                 return false;