+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.2.4
+-----
+- Fixed product attribute regex
+- Fixed product specific pull mapping
+- Fixed product pseudo and base price reset
+- Fixed category sorting for different languages
+- Added product variation and variation value sorting
+- Changed product attribute start value from 3 to 4
+
 1.2.3
 -----
 - Changed product image main and position handling